+++ Seashore/source/tool/PositionTool.m
@@ -119,7 +119,7 @@
 	int deltax = where.x - initialPoint.x, deltay = where.y - initialPoint.y;
 	IntPoint oldOffsets = {0};
 	NSPoint activeCenter = NSMakePoint([activeLayer xoff] + [(SeaLayer *)activeLayer width] / 2, [activeLayer yoff] + [(SeaLayer *)activeLayer height] / 2);
-	float original, current;
+	CGFloat original, current;
 	
 	// Vary behaviour based on function
 	switch ([options  toolFunction]) {
